# docs/openapi.yaml
openapi: 3.0.0
info:
  title: InChat API
  description: InChat 应用的 API 文档
  version: 1.0.0
  contact:
    name: Your Name
    email: your.email@example.com
servers:
  - url: http://localhost:3006
    description: 开发服务器
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/users/{userId}/status:
    get:
      tags:
        - 用户
      summary: 获取用户状态
      description: 获取指定用户的状态信息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 用户ID
      responses:
        '200':
          description: 成功获取用户状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [online, offline, away, busy]
                    example: "online"
                  customStatus:
                    type: string
                    example: "正在开会"
                  lastSeen:
                    type: string
                    format: date-time
                    example: "2024-03-15T12:00:00Z"
                  isTypingToYou:
                    type: boolean
                    example: false
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "用户不存在"